backend/requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=1.4.0,<2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Document Processing
pdfplumber>=0.9.0
python-magic>=0.4.27

# Vector Database
pgvector>=0.2.0
sqlalchemy-utils>=0.41.0

# AI and Embeddings
sentence-transformers>=2.2.2
langchain>=0.0.267
openai>=0.27.8
